Item getestes und kauf getestet

+++ Sybau_Backend/Sybau_Backend/Models/Item.cs
@@ -9,7 +9,7 @@ public class Item : BaseEntity<int>
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
 
 
-    public Item(string name, string description, ItemType type, double price, int xpBoostPercent)
+    public Item(string name, string description, ItemType type, int price, int xpBoostPercent)
     {
         if(string.IsNullOrWhiteSpace(name)) throw new ArgumentNullException(nameof(name));
         if(string.IsNullOrWhiteSpace(description)) throw new ArgumentNullException(nameof(description));
@@ -27,7 +27,7 @@ public class Item : BaseEntity<int>
     public string Description { get; set; }
     
     public ItemType Type { get; set; }
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int XpBoostPercent { get; set; }
     public ICollection<UserItem> UserItems { get; set; } = new List<UserItem>();
 }
